adding forest node to assertions, and cleaning up main.go by moving all the functions to workload to cleanup and make code readable
