+++ workload/resources/config.json
@@ -22,7 +22,7 @@
     },
     {
       "name":"Forest",
-      "rpcurl": "http://10.20.20.28:23456/rpc/v1",
+      "rpcurl": "http://10.20.20.28:3456/rpc/v1",
       "authtokenpath":"/root/devgen/forest/jwt"
     }
   ],
