+++ workload/resources/consensus.go
@@ -336,3 +336,179 @@ func (cc *ConsensusChecker) CheckConsensus(ctx context.Context, height abi.Chain
 		return nil
 	}, "Consensus check operation")
 }
+
+// PerformConsensusCheck checks consensus between nodes
+func PerformConsensusCheck(ctx context.Context, config *Config, height int64) error {
+	log.Printf("[INFO] Starting consensus check...")
+
+	checker, err := NewConsensusChecker(ctx, config.Nodes)
+	if err != nil {
+		log.Printf("[ERROR] Failed to create consensus checker: %v", err)
+		return nil
+	}
+
+	// If height is 0, we'll let the checker pick a random height
+	if height == 0 {
+		log.Printf("[INFO] No specific height provided, will check consensus at a random height")
+	} else {
+		log.Printf("[INFO] Will check consensus starting at height %d", height)
+	}
+
+	// Run the consensus check
+	err = checker.CheckConsensus(ctx, abi.ChainEpoch(height))
+	if err != nil {
+		log.Printf("[WARN] Consensus check failed, will retry: %v", err)
+		return err // Return original error for retry
+	}
+
+	log.Printf("[INFO] Consensus check completed successfully")
+	return nil
+}
+
+// PerformSendConsensusFault sends a consensus fault
+func PerformSendConsensusFault(ctx context.Context) error {
+	log.Println("[INFO] Attempting to send a consensus fault...")
+	err := SendConsensusFault(ctx)
+	if err != nil {
+		log.Printf("[ERROR] Failed to send consensus fault: %v", err)
+		return nil
+	}
+	log.Println("[INFO] SendConsensusFault operation initiated. Check further logs for details.")
+	return nil
+}
+
+// PerformCheckFinalizedTipsets checks finalized tipsets
+func PerformCheckFinalizedTipsets(ctx context.Context) error {
+	log.Printf("[INFO] Starting finalized tipset comparison...")
+
+	// Load configuration
+	config, err := LoadConfig("/opt/antithesis/resources/config.json")
+	if err != nil {
+		log.Printf("[ERROR] Failed to load config: %v", err)
+		return nil
+	}
+
+	// Filter nodes to get V1 and V2 nodes separately
+	v1Nodes := FilterLotusNodesV1(config.Nodes)
+	v2Nodes := FilterLotusNodesWithV2(config.Nodes)
+
+	if len(v1Nodes) < 2 {
+		log.Printf("[ERROR] Need at least two Lotus V1 nodes for this test, found %d", len(v1Nodes))
+		return nil
+	}
+	if len(v2Nodes) < 2 {
+		log.Printf("[ERROR] Need at least two Lotus V2 nodes for this test, found %d", len(v2Nodes))
+		return nil
+	}
+
+	// Connect to V1 nodes to get chain heads and find common height range
+	api1, closer1, err := ConnectToNode(ctx, v1Nodes[0])
+	if err != nil {
+		log.Printf("[ERROR] Failed to connect to %s: %v", v1Nodes[0].Name, err)
+		return nil
+	}
+	defer closer1()
+
+	api2, closer2, err := ConnectToNode(ctx, v1Nodes[1])
+	if err != nil {
+		log.Printf("[ERROR] Failed to connect to %s: %v", v1Nodes[1].Name, err)
+		return nil
+	}
+	defer closer2()
+
+	ch1, err := api1.ChainHead(ctx)
+	if err != nil {
+		log.Printf("[ERROR] Failed to get chain head from %s: %v", v1Nodes[0].Name, err)
+		return nil
+	}
+
+	ch2, err := api2.ChainHead(ctx)
+	if err != nil {
+		log.Printf("[ERROR] Failed to get chain head from %s: %v", v1Nodes[1].Name, err)
+		return nil
+	}
+
+	h1 := ch1.Height()
+	h2 := ch2.Height()
+
+	log.Printf("[INFO] Current height %d for node %s", h1, v1Nodes[0].Name)
+	log.Printf("[INFO] Current height %d for node %s", h2, v1Nodes[1].Name)
+
+	// Find the common height between both nodes
+	var commonHeight int64
+	if h1 < h2 {
+		commonHeight = int64(h1)
+	} else {
+		commonHeight = int64(h2)
+	}
+
+	// Ensure we have enough history for F3 finalized tipset comparison
+	// F3 starts from epoch 20, so we need at least 30 epochs to have a meaningful range
+	if commonHeight < 30 {
+		log.Printf("[WARN] chain height too low for finalized tipset comparison (common: %d, required: 30 for F3 range)", commonHeight)
+		return nil
+	}
+
+	// Select a random height within the F3 range
+	// F3 starts from epoch 20, and we leave 10 epochs buffer from the tip
+	rand.Seed(time.Now().UnixNano())
+	maxHeight := commonHeight - 10 // Leave 10 epochs buffer from tip
+	minHeight := int64(20)         // F3 starts from epoch 20
+
+	if maxHeight <= minHeight {
+		log.Printf("[WARN] Not enough height range for finalized tipset comparison (min: %d, max: %d)", minHeight, maxHeight)
+		return nil
+	}
+
+	randomHeight := minHeight + rand.Int63n(maxHeight-minHeight+1)
+	log.Printf("[INFO] Selected height %d for finalized tipset comparison (range: %d-%d)", randomHeight, minHeight, maxHeight)
+
+	// Connect to V2 nodes for finalized tipset comparison
+	api11, closer11, err := ConnectToNodeV2(ctx, v2Nodes[0])
+	if err != nil {
+		log.Printf("[ERROR] Failed to connect to %s: %v", v2Nodes[0].Name, err)
+		return nil
+	}
+	defer closer11()
+
+	api22, closer22, err := ConnectToNodeV2(ctx, v2Nodes[1])
+	if err != nil {
+		log.Printf("[ERROR] Failed to connect to %s: %v", v2Nodes[1].Name, err)
+		return nil
+	}
+	defer closer22()
+
+	// Chain walk: Check 10 tipsets down from the selected height
+	log.Printf("[INFO] Starting chain walk from height %d down to %d", randomHeight, randomHeight-9)
+
+	for i := randomHeight; i >= randomHeight-9; i-- {
+		log.Printf("[INFO] Checking finalized tipset at height %d", i)
+		heightSelector := types.TipSetSelectors.Height(abi.ChainEpoch(i), true, types.TipSetAnchors.Finalized)
+
+		ts1, err := api11.ChainGetTipSet(ctx, heightSelector)
+		if err != nil {
+			log.Printf("failed to get finalized tipset by height from %s: %v", v2Nodes[0].Name, err)
+			return nil
+		}
+		log.Printf("[INFO] Finalized tipset %s on %s at height %d", ts1.Cids(), v2Nodes[0].Name, i)
+
+		ts2, err := api22.ChainGetTipSet(ctx, heightSelector)
+		if err != nil {
+			log.Printf("failed to get finalized tipset by height from %s: %v", v2Nodes[1].Name, err)
+			return nil
+		}
+		log.Printf("[INFO] Finalized tipset %s on %s at height %d", ts2.Cids(), v2Nodes[1].Name, i)
+
+		assert.Always(ts1.Equals(ts2), "Chain synchronization test: Finalized tipset should always match",
+			map[string]interface{}{
+				"requirement": "Chain synchronization",
+				"ts1":         ts1.Cids(),
+				"ts2":         ts2.Cids(),
+			})
+
+		log.Printf("[INFO] Finalized tipsets %s match successfully at height %d", ts1.Cids(), i)
+	}
+
+	log.Printf("[INFO] Chain walk completed successfully - all 10 finalized tipsets match between nodes")
+	return nil
+}
